Clarify some descriptions, remove redundant info
The headers aren't blank because the user has to supply them with
values. There's no space to allocate because they're just headers.

Verifying magic numbers is something every program should do
anyway to have at least the slightest guarantee the file it's using
is actually in the expected format, the user doesn't really need to
know that.